Patch from boc & me for HPUX compilation and colormap issues
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@4844 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/text/fmt/xp/fp_Column.cpp
@@ -427,7 +427,7 @@ UT_uint32 fp_Container::distanceFromPoint(UT_sint32 x, UT_sint32 y)
 		return dx;
 	}
 
-	UT_uint32 dist = (UT_uint32) (sqrt((dx * dx) + (dy * dy)));
+	UT_uint32 dist = (UT_uint32) (sqrt((float)(dx * dx) + (dy * dy)));
 
 	UT_ASSERT(dist > 0);
